+++ app/specific/Main.js
@@ -25,10 +25,10 @@ var Main_isReleased = false;
 var Main_isDebug = false;
 
 var Main_stringVersion = '3.0';
-var Main_stringVersion_Min = '.240';
-var Main_version_java = 31;//Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
+var Main_stringVersion_Min = '.241';
+var Main_version_java = 32;//Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
 var Main_minversion = 'August 15 2020';
-var Main_version_web = 51;//Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
+var Main_version_web = 52;//Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
 var Main_versionTag = Main_stringVersion + Main_stringVersion_Min + '-' + Main_minversion;
 
 var Main_cursorYAddFocus = -1;
@@ -620,8 +620,11 @@ function Main_Changelog() {
 
     var changelogObj = [
         {
-            title: "Web Version August 15 2020",
-            changes: ["Mitigate the freeze effects when using accessibility service, most noticeably when exiting the player, slowdowns do to accessibility service can't be fixed only mitigated, read more about it https://tinyurl.com/applag"]
+            title: "Apk Version 3.0.241 - Web Version August 15 2020",
+            changes: [
+                "Mitigate the freeze effects when using accessibility service, most noticeably when exiting the player, slowdowns do to accessibility service can't be fixed only mitigated, read more about it https://tinyurl.com/applag",
+                "General performance improves and bug fixes"
+            ]
         },
         {
             title: "Apk Version 3.0.240 - August 10 2020",